Set serial at zone buildout.
Since we update the zones in bulk, we might just as well specify a single global serial at zone creation.
Also: keep the serial a number, makes more sense to increment it.
Todo: save the last serial somewhere (local variable in the provisioning database? special record?) so that it can be consistent at startup between servers.